+++ extensions/src/extensions/getSelectedIds.ts
@@ -1,0 +1,14 @@
+import {GridFunction, GridPluginType, RowSelector} from '@anglr/grid';
+
+/**
+ * Gets currently selected ids
+ */
+export function getSelectedIds<TId>(): GridFunction<TId[]>
+{
+    return grid =>
+    {
+        const rowSelector = grid.getPlugin<RowSelector<unknown, unknown, TId>>(GridPluginType.RowSelector);
+
+        return rowSelector.selectedIds();
+    };
+}